backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
supabase>=2.4.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
websockets>=13.0,<16
